fix(duplicate): handle multi-selection and groups in Cmd+D
The keyboard duplicate handler only operated on the first selected
target. Multi-selection of mixed items duplicated just one entity,
diverging from copy/paste which preserves the whole selection.

Extracted a shared duplicateSelection() routine and routed both the
keyboard binding and canvas-duplicate-selection IPC through it. Single
selection now also dispatches groups to duplicateGroup; previously a
single group + Cmd+D threw from duplicateEntity.
